add pets, terrestrial, spider and fish classes

+++ src/main/java/org/example/Terrestrial.java
@@ -1,9 +1,14 @@
 package org.example;
 
-public class Terrestrial extends Pets{
-    public Terrestrial(int numberOfLegs, String color, boolean isStinky) {
+import java.math.BigDecimal;
+
+public class Terrestrial extends Pets {
+    public Terrestrial(BigDecimal numberOfLegs, String color, boolean isStinky) {
         super(numberOfLegs, color, isStinky);
+
         this.setStinky(false);
         this.setColor("");
+
+
     }
 }
